fix: GM invisibility preserves body/head, smoother pulsing
Server: Don't zero body/head when going invisible. Instead keep
them intact, send BP (remove) to others, SetInvisible to self.
When turning visible, broadcast CC to restore on others' screens.

Client: TransparenciaBody pulsing is now time-based (~4s cycle)
instead of frame-based, for a smoother fade effect. Changed field
from int to float for smooth interpolation.

+++ client/Scripts/Game/Character.cs
@@ -38,7 +38,7 @@ public class Character
     public bool Levitating;
 
     // VB6: dead character transparency pulsing (TransparenciaBody oscillates 0-100)
-    public int TransparenciaBody;  // 0-100, alpha = this + 45
+    public float TransparenciaBody;  // 0-100, alpha = this + 45
     public bool Llegoalatransp;    // false=increasing, true=decreasing
 
     // VB6: EmoticonLoops countdown (separate from FX slots)
